refactor: move handler to index
[no ci]
